+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-07-23 11:54:49
-// Commit: feat(i18n): add async i18n stream provider
+// Updated at: 2026-07-23 12:21:49
+// Commit: refactor(i18n): extract reusable atomic widgets from i18n
 
 class i18nComponentState {
-  final String commitRef = '3056';
-  final String lastMessage = 'feat(i18n): add async i18n stream provider';
-  final DateTime timestamp = DateTime.parse('2026-07-23T11:54:49+05:00');
+  final String commitRef = '3058';
+  final String lastMessage = 'refactor(i18n): extract reusable atomic widgets from i18n';
+  final DateTime timestamp = DateTime.parse('2026-07-23T12:21:49+05:00');
 }